+++ .gitignore
@@ -5,7 +5,6 @@ __pycache__/
 
 # C extensions
 *.so
-.pre-commit-config.yaml
 # Distribution / packaging
 .Python
 build/
