+++ python/xorbits/deploy/cluster/slurm.py
@@ -1,12 +1,22 @@
+# Copyright 2022-2023 XProbe Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 import atexit
 import logging
 import os
 import re
 import subprocess
 import time
-
-import xorbits
-import xorbits.numpy as np
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -188,17 +198,3 @@ class SLURMCluster:
                     time.sleep(sleep_interval)
             except Exception as e:
                 logger.error(str(e))
-
-
-if __name__ == "__main__":
-    exp = SLURMCluster(
-        job_name="xorbits",
-        num_nodes=2,
-        output_dir="/shared_space/output.out",
-        time="00:30:00",
-    )
-    address = exp.run()
-    logger.info(address)
-    xorbits.init(address)
-    test = np.random.rand(100, 100).mean()
-    logging.info(f"test_result_{test}")
